+++ dataaccessobject/rawdb/accessor_beacon.go
