+++ dataaccessobject/rawdb/accessor_burn.go
