+++ dataaccessobject/rawdb/accessor_pde.go
