+++ dataaccessobject/rawdb/error.go
